More Crud

+++ SacramentMeetingPlanner/Controllers/AssignmentsController.cs
@@ -72,13 +72,23 @@ namespace SacramentMeetingPlanner.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,MemberID,MeetingID,assignment,Topic")] Assignment assignments)
+        public async Task<IActionResult> Create([Bind("MemberID,MeetingID,assignment,Topic")] Assignment assignments)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(assignments);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(assignments);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(assignments);
         }
@@ -102,40 +112,38 @@ namespace SacramentMeetingPlanner.Controllers
         // POST: Assignments/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,MemberID,MeetingID,assignment,Topic")] Assignment assignments)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != assignments.ID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var assignmentToUpdate = await _context.Assignments.FirstOrDefaultAsync(a => a.ID == id);
+            if (await TryUpdateModelAsync<Assignment>(
+                assignmentToUpdate,
+                "",
+                a => a.MemberID, a => a.MeetingID, a => a.assignment, a => a.Topic))
             {
                 try
                 {
-                    _context.Update(assignments);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException)
                 {
-                    if (!AssignmentExists(assignments.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists, " +
+                "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(assignments);
+            return View(assignmentToUpdate);
         }
 
         // GET: Assignments/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -143,10 +151,18 @@ namespace SacramentMeetingPlanner.Controllers
             }
 
             var assignment = await _context.Assignments
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (assignment == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                 "Delete failed. Try again, and if the problem persists " +
+                 "see your system administrator.";
             }
 
             return View(assignment);
@@ -158,9 +174,22 @@ namespace SacramentMeetingPlanner.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var assignment = await _context.Assignments.FindAsync(id);
-            _context.Assignments.Remove(assignment);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (assignment == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Assignments.Remove(assignment);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool AssignmentExists(int id)
